Add dependencies and fix project name
