+++ app/src/androidTest/java/fr/im/salimi/projectmanager/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package fr.im.salimi.teammanager
+package fr.im.salimi.projectmanager
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
