+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="fr.im.salimi.teammanager">
+    package="fr.im.salimi.projectmanager">
 
     <application
         android:allowBackup="true"
@@ -10,7 +10,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.TeamManager">
         <activity
-            android:name=".MainActivity"
+            android:name="fr.im.salimi.projectmanager.MainActivity"
             android:label="@string/app_name"
             android:theme="@style/Theme.TeamManager.NoActionBar">
             <intent-filter>
